Restruct code
