+++ src/Components/App.js
@@ -1,7 +1,7 @@
 import React,{Component, Fragment} from 'react';
 import Game from './Game';
 import {Header, Footer} from './Layouts'
-import axios from 'axios';
+
 const API_URL = 'http://starnavi-frontend-test-task.herokuapp.com';
 export default class extends Component {
