+++ src/Components/Game/ControlPanel/DifficultyMenu.jsx
@@ -1,9 +1,10 @@
-import React from 'react';
+import React, {useEffect} from 'react';
 import { makeStyles } from '@material-ui/core/styles'
 import FormControl from '@material-ui/core/FormControl';
 import InputLabel from '@material-ui/core/InputLabel';
 import Select from '@material-ui/core/Select';
-import MenuItem from '@material-ui/core/MenuItem'
+import MenuItem from '@material-ui/core/MenuItem';
+import axios from 'axios';
 
 const useStyles = makeStyles(theme => ({
 
@@ -19,6 +20,16 @@ const useStyles = makeStyles(theme => ({
 
 export default props => {
   const classes = useStyles();
+
+  useEffect(()=>{ axios.get(`${props.api}/game-settings`)
+      .then(res => {
+        const gameModePresets = res.data;
+        // setLoading(false);
+      })
+    }, [])
+
+
+
   const [values, setValues] = React.useState({
      age: '',
      name: 'hai',
